Interceptor: U-wing can be combined with bottom carriage
And by extension bottom rear and full X-wing

U-wing is no longer part of group 11, since it
can be combined with rear/full X-wings.

For simplicities sake the U-fin is also in group 32 with
the U-wing. This means we don't have to add separate okGroup
for the U-wing fin for everything that's compatible with uwing.

Group 31 is gone (Was U-wing fin)